subcmd/apply: Support fail-safe approach
Up until now, the `apply` command implemented a fail-fast
strategy. This means that an attempt to configure a batch
of topics (by passing a path to a folder contains a list
of topics configurations files) will be terminated upon
the first failure (w/o even trying to apply the configuration
for the next-in-line topics).
This commit introduces the fail-safe support:
It allows the user to run the `apply` command in a fail-safe
manner, which means that instead of failing upon the first
failure, we'll simply continue to the next topic configuration,
while aggregating all the errors along the way.
Eventually, if there were any errors, all of them will be shown.

Signed-off-by: shimon-armis <shimon.turjeman@armis.com>