Set up Dockerfile and scripts (#9)
This prepares the scripts to wire up topicctl with the python logic that
parses the json output and notifies us.

For now there is one simple single file python script so I left it in the
py directory together with the requirements.txt file.
If we start having more scripts (likely) we will build a python project
with tests and a pyproject.toml in the py directory.

I replaced ENTRY_POINT with A CMD entry in the Dockerfile. This allows us
to override the behavior and both run a simple topicctl and a topicctl
with notifications.